Rename API Supervision to Logbook Operation
