+++ deployment/scripts/mongod/1.0.0/207_iam_ref.js.j2
@@ -408,25 +408,25 @@ db.profiles.insert({
 });
 
 db.profiles.insert({
-  "_id": "system_supervision_api_profile",
+  "_id": "system_logbook_operation_profile",
   "identifier" : NumberInt(maxIdProfile++),
-  "name": "Supervision API Profile",
-  "description": "Supervision Profile api",
+  "name": "Journal des Opérations",
+  "description": "Logbook Operation Profile api",
   "tenantIdentifier": NumberInt({{ vitamui_platform_informations.proof_tenant }}),
-  "applicationName": "SUPERVISION_API_APP",
+  "applicationName": "LOGBOOK_OPERATION_APP",
   "level": "",
   "enabled": true,
   "readonly": false,
   "customerId": "system_customer",
   "roles": [
   {
-    "name": "ROLE_CREATE_SUPERVISION_API"
+    "name": "ROLE_CREATE_LOGBOOK_OPERATION_API"
   },
   {
-    "name": "ROLE_GET_SUPERVISION_API"
+    "name": "ROLE_GET_LOGBOOK_OPERATION_API"
   },
   {
-    "name": "ROLE_GET_ALL_SUPERVISION_API"
+    "name": "ROLE_GET_ALL_LOGBOOK_OPERATION_API"
   },
   {
        "name": "ROLE_LOGBOOKS"
@@ -464,7 +464,7 @@ db.groups.updateOne( {
                 "system_secure",
                 "system_dsl",
                 "system_probative_value",
-                "system_supervision_api_profile"
+                "system_logbook_operation_profile"
         	]
         }
     }
