+++ deployment/scripts/mongod/1.0.0/210_security_ref.js.j2
@@ -25,7 +25,7 @@ db.contexts.insert({
         "ROLE_GET_CONTEXTS", "ROLE_UPDATE_CONTEXTS", "ROLE_CREATE_CONTEXTS",
         "ROLE_GET_SECURITY_PROFILES", "ROLE_UPDATE_SECURITY_PROFILES", "ROLE_CREATE_SECURITY_PROFILES", "ROLE_DELETE_SECURITY_PROFILES",
         "ROLE_GET_OPERATIONS", "ROLE_RUN_AUDITS", "ROLE_RUN_PROBATIVE_VALUE",
-        "ROLE_CREATE_SUPERVISION_API", "ROLE_GET_SUPERVISION_API", "ROLE_GET_ALL_SUPERVISION_API", "ROLE_LOGBOOKS"
+        "ROLE_CREATE_LOGBOOK_OPERATION_API", "ROLE_GET_LOGBOOK_OPERATION_API", "ROLE_GET_ALL_LOGBOOK_OPERATION_API", "ROLE_LOGBOOKS"
 
     ]
 });
