+++ ui/ui-frontend/projects/referential/src/app/logbook-operation/event-filter.interface.ts
