+++ ui/ui-frontend/projects/referential/src/app/logbook-operation/logbook-operation-detail/logbook-operation-detail.component.html
@@ -8,7 +8,7 @@
     <div class="header-container">
       <div class="title">
         <div class="status-badge-mini" [ngClass]="event | eventTypeBadgeClass">
-          <i class="vitamui-icon vitamui-icon-box-hand no-margin"></i>
+          <i class="vitamui-icon vitamui-icon-api-supervision vitamui-row-icon status-badge"></i>
         </div>
         <h2>{{ event?.id }}</h2>
       </div>
