+++ ui/ui-frontend/projects/referential/src/app/logbook-operation/logbook-operation-list/event-type-badge-class.pipe.ts
