+++ ui/ui-frontend/projects/referential/src/app/logbook-operation/logbook-operation-list/event-type-color-class.pipe.spec.ts
