+++ ui/ui-frontend/projects/referential/src/app/logbook-operation/logbook-operation-list/last-event.pipe.spec.ts
