+++ ui/ui-frontend/projects/referential/src/app/logbook-operation/logbook-operation-list/logbook-operation-list.component.html
@@ -48,7 +48,7 @@
       <td style="width:60px;" class="no-first-cell-padding">
         <div class="vitamui-cell-icon aligned">
           <div class="status-badge" [ngClass]="event | eventTypeBadgeClass">
-            <i class="vitamui-icon vitamui-icon-ingest vitamui-row-icon status-badge"></i>
+            <i class="vitamui-icon vitamui-icon-api-supervision vitamui-row-icon status-badge"></i>
           </div>
         </div>
       </td>
