+++ ui/ui-frontend/projects/referential/src/app/logbook-operation/logbook-operation-list/logbook-operation-list.component.spec.ts
@@ -34,54 +34,39 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-import { DEFAULT_PAGE_SIZE, Direction, Event, InfiniteScrollTable, PageRequest } from 'ui-frontend-common';
+import { EMPTY } from 'rxjs';
 
-import {
-  Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges
-} from '@angular/core';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { EventFilter } from '../event-filter.interface';
 import { LogbookSearchService } from '../logbook-search.service';
+import { EventTypeBadgeClassPipe } from './event-type-badge-class.pipe';
+import { EventTypeColorClassPipe } from './event-type-color-class.pipe';
+import { LastEventPipe } from './last-event.pipe';
+import { LogbookOperationListComponent } from './logbook-operation-list.component';
 
-@Component({
-  selector: 'app-api-supervision-list',
-  templateUrl: './api-supervision-list.component.html',
-  styleUrls: ['./api-supervision-list.component.scss']
-})
-export class ApiSupervisionListComponent extends InfiniteScrollTable<Event> implements OnInit, OnChanges {
+describe('LogbookOperationListComponent', () => {
+  let component: LogbookOperationListComponent;
+  let fixture: ComponentFixture<LogbookOperationListComponent>;
 
-  @Input() tenantIdentifier: number;
-  @Input() filters: Readonly<EventFilter>;
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [LogbookOperationListComponent, LastEventPipe, EventTypeColorClassPipe, EventTypeBadgeClassPipe ],
+      providers: [
+        { provide: LogbookSearchService, useValue: { search: () => EMPTY } },
+      ],
+      schemas: [NO_ERRORS_SCHEMA]
+    })
+    .compileComponents();
+  }));
 
-  @Output() eventClick = new EventEmitter<Event>();
+  beforeEach(() => {
+    fixture = TestBed.createComponent(LogbookOperationListComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
 
-  orderBy = 'name';
-  direction = Direction.ASCENDANT;
-
-  constructor(public logbookSearchService: LogbookSearchService) {
-    super(logbookSearchService);
-  }
-
-  ngOnInit() {
-    this.refreshList();
-  }
-
-  ngOnChanges(changes: SimpleChanges) {
-    if (changes.tenantIdentifier || changes.filters) {
-      this.refreshList();
-    }
-  }
-
-  refreshList() {
-    const pageRequest = new PageRequest(0, DEFAULT_PAGE_SIZE, this.orderBy, this.direction);
-
-    const query = JSON.stringify(LogbookSearchService.buildVitamQuery(pageRequest, this.filters));
-
-    this.search(new PageRequest(0, DEFAULT_PAGE_SIZE, this.orderBy, this.direction, query));
-  }
-
-  selectEvent(event: Event) {
-    this.eventClick.emit(event);
-  }
-
-}
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
