+++ ui/ui-frontend/projects/referential/src/app/logbook-operation/logbook-operation.component.scss
