+++ ui/ui-frontend/projects/referential/src/app/logbook-operation/logbook-search.service.spec.ts
