+++ ui/ui-frontend/projects/referential/src/app/logbook-operation/logbook-search.service.ts
@@ -34,10 +34,10 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
+import * as moment from 'moment';
 import {
   Event, LogbookApiService, PageRequest, SearchService, VitamSelectOperator, VitamSelectQuery
 } from 'ui-frontend-common';
-import * as moment from 'moment';
 
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
